requirements.txt
requests==2.25.1
Pillow==8.1.0
python-dotenv==0.15.0
flickrapi==2.4.0